requirements.txt
boto3==1.20.31
dj_database_url == 0.5.0
Django==3.2.10
DjangoRestFramework==3.13.1
django-allauth==0.47.0
django-cors-headers==3.3.0
django-daterange-filter==1.3.0
django-filter==21.1
django-materializecss-form==1.1.12
dj-rest-auth==1.1.0
django-storages==1.12.3
django-webpack-loader==0.6.0
gunicorn==19.9.0
numpy==1.21.3
pandas==1.3.5
Pillow==9.1.1
psycopg2-binary==2.9.3